Created decimal() and buttonDecimal
Began working on decimals and attempting to allow decimals in
calculators (ex. 1.3+4.5)

I added a decimal button next to the zero, it may get moved over but
not sure about changing much in terms of the design yet.

I defined the decimal() function and have began looking into ways of
allowing it to work. Looking into having it take the operationOne/Two
and setting it to operationeOne/Two as a float while still editing so
if they enter other things it will add after the decimal point.