Restart work on decimal()
Was trying to do stuff but because of how the strings and numbers are
put together I might have to re-write a decent amount of the program
and have one variable that does it (workspace variable) so I can modify
the actual workspace variable rather than modifying the operations as
much. Might wait and put that in the ToDo for 3.0 and work on fixing
the bugs that are currently existing.