About Button + Imports
Added an about button that executes the about() function that will
display information about the project that can be defined in the Meta
Data section at the top of the program. Imported tKMessageBox to allow
a popup when the function is executed.

Added the new import to the imports section in the readme.